rename setup.py -> build.py
